kpitechnicaltest/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
django-extensions==3.1.3
django-heroku==0.3.1
gunicorn==20.1.0
pip==21.3.1
psycopg2==2.9.3
pytz==2021.3
setuptools==57.4.0
sqlparse==0.4.2
whitenoise==5.3.0
djangorestframework
django-filter
dj_rest_auth
django-cors-headers